- Refactor reducers to all have -action-creators file. - None of them use reducerId.
