+++ packages/redux/src/features/higher-order-reducers/record-object-reducer/__tests__/record-object-reducer.test.ts
@@ -1,26 +1,29 @@
-import { recordObjectReducer } from "../record-object-reducer";
-import { createEntityReducer } from "../../../reducer-factories-with-reducer-id/entity-reducer/entity-reducer";
+import {
+  recordObjectAction,
+  recordObjectReducer
+} from "../record-object-reducer";
+import { createEntityReducer } from "../../../reducer-factories/entity-reducer/entity-reducer";
+import { createEntityActions } from "../../../reducer-factories/entity-reducer/entity-actions";
+
+type Entity = {
+  id: string;
+  email: string;
+};
 
 describe("record-object-reducer", () => {
   describe("recordObjectReducer", () => {
     it("runs the reducer on a field on the record", () => {
+      const actions = createEntityActions<Entity>();
       const reducer = recordObjectReducer(
-        createEntityReducer<{
-          id: string;
-          email: string;
-        }>("test", { id: "", email: "" })
+        createEntityReducer<Entity>({ id: "", email: "" })
       );
 
       const r = reducer(
         {},
-        {
-          recordId: "id1",
-          action: {
-            type: "ENTITY:SET_ENTITY",
-            reducerId: "test",
-            entity: { id: "id1", email: "johan" }
-          }
-        }
+        recordObjectAction(
+          "id1",
+          actions.setEntity({ id: "id1", email: "johan" })
+        )
       );
 
       expect(r.id1.entity.email).toBe("johan");
