--- packages/redux/src/features/reducer-factories-with-reducer-id/sort-order-by-id-reducer/sort-order-by-id-reducer.ts
@@ -1,34 +1,0 @@
-import { SortOrderByIdAction } from "./sort-order-by-id-actions";
-
-export interface SortOrderByIdState {
-  ids: Array<string> | undefined;
-}
-
-const INITIAL_STATE: SortOrderByIdState = {
-  ids: undefined
-};
-
-export const createSortOrderByIdReducer = (reducerId: string) => (
-  state: SortOrderByIdState = INITIAL_STATE,
-  action: SortOrderByIdAction
-): SortOrderByIdState => {
-  if (action.reducerId !== reducerId) {
-    return state;
-  }
-  switch (action.type) {
-    case "SORT_ORDER_BY_ID:SET_SORT_ORDER":
-      return {
-        ...state,
-        ids: action.ids
-      };
-
-    case "SORT_ORDER_BY_ID:CLEAR_SORT_ORDER":
-      return {
-        ...state,
-        ids: undefined
-      };
-
-    default:
-      return state;
-  }
-};
