+++ packages/redux/src/features/reducer-factories/entity-reducer/entity-actions.ts
@@ -1,0 +1,13 @@
+export type EntityAction<T> =
+  | EntitySetEntityAction<T>
+  | EntitySetEntityFieldsAction<T>;
+
+export interface EntitySetEntityAction<T> {
+  type: "ENTITY:SET_ENTITY";
+  entity: T;
+}
+
+export interface EntitySetEntityFieldsAction<T> {
+  type: "ENTITY:SET_ENTITY_FIELDS";
+  fields: Partial<T>;
+}
