+++ packages/redux/src/features/reducer-factories/selected-ids-reducer/selected-ids-reducer.ts
@@ -1,0 +1,39 @@
+import { Reducer } from "redux";
+import { SelectedIdsAction } from "./selected-ids-actions";
+
+export interface SelectedIdsState {
+  selectedIds: Array<string>;
+}
+
+export const createSelectedIdsReducerInitialState = (): SelectedIdsState => ({
+  selectedIds: []
+});
+
+const INITIAL_STATE = createSelectedIdsReducerInitialState();
+
+export type SelectedIdsReducer = Reducer<SelectedIdsState, SelectedIdsAction>;
+
+export const createSelectedIdsReducer = (): SelectedIdsReducer => (
+  state = INITIAL_STATE,
+  action
+) => {
+  switch (action.type) {
+    case "SELECTED_IDS:SET_SELECTED_IDS": {
+      const { selectedIds } = action;
+      return {
+        ...state,
+        selectedIds
+      };
+    }
+    case "SELECTED_IDS:CLEAR_SELECTED_IDS": {
+      return {
+        ...state,
+        selectedIds: []
+      };
+    }
+    default:
+      return state;
+  }
+};
+
+export const selectedIdsReducer = createSelectedIdsReducer();
