+++ packages/redux/src/features/reducer-factories/sort-order-by-id-reducer/sort-order-by-id-action-creators.ts
@@ -1,0 +1,23 @@
+import {
+  SortOrderByIdClearSortOrderAction,
+  SortOrderByIdSetSortOrderAction
+} from "./sort-order-by-id-actions";
+
+export interface SortOrderByIdActions {
+  setSortOrder: (ids: Array<string>) => SortOrderByIdSetSortOrderAction;
+  clearSortOrder: () => SortOrderByIdClearSortOrderAction;
+}
+
+export const createSortOrderByIdActions = (): SortOrderByIdActions => {
+  return {
+    setSortOrder: ids => ({
+      ids,
+      type: "SORT_ORDER_BY_ID:SET_SORT_ORDER"
+    }),
+    clearSortOrder: () => ({
+      type: "SORT_ORDER_BY_ID:CLEAR_SORT_ORDER"
+    })
+  };
+};
+
+export const sortOrderByIdActions = createSortOrderByIdActions();
