+++ packages/redux/src/features/reducer-factories/sort-order-by-id-reducer/sort-order-by-id-actions.ts
@@ -1,0 +1,12 @@
+export type SortOrderByIdAction =
+  | SortOrderByIdSetSortOrderAction
+  | SortOrderByIdClearSortOrderAction;
+
+export interface SortOrderByIdSetSortOrderAction {
+  type: "SORT_ORDER_BY_ID:SET_SORT_ORDER";
+  ids: Array<string>;
+}
+
+export interface SortOrderByIdClearSortOrderAction {
+  type: "SORT_ORDER_BY_ID:CLEAR_SORT_ORDER";
+}
