+++ packages/redux/src/index.ts
@@ -14,26 +14,31 @@ export * from "./features/reducer-factories/entity-list-reducer/entity-list-acti
 export * from "./features/reducer-factories/entity-list-reducer/entity-list-reducer";
 export * from "./features/reducer-factories/entity-list-reducer/entity-list-selectors";
 
-export * from "./features/reducer-factories-with-reducer-id/entity-reducer/entity-actions";
-export * from "./features/reducer-factories-with-reducer-id/entity-reducer/entity-reducer";
-export * from "./features/reducer-factories-with-reducer-id/entity-reducer/entity-selectors";
+export * from "./features/reducer-factories/entity-by-id-reducer/entity-by-id-reducer";
+export * from "./features/reducer-factories/entity-by-id-reducer/entity-by-id-actions";
+export * from "./features/reducer-factories/entity-by-id-reducer/entity-by-id-action-creators";
+export * from "./features/reducer-factories/entity-by-id-reducer/entity-by-id-selectors";
 
-export * from "./features/reducer-factories-with-reducer-id/entity-by-id-reducer/entity-by-id-reducer";
-export * from "./features/reducer-factories-with-reducer-id/entity-by-id-reducer/entity-by-id-actions";
-export * from "./features/reducer-factories-with-reducer-id/entity-by-id-reducer/entity-by-id-selectors";
+export * from "./features/reducer-factories/entity-reducer/entity-reducer";
+export * from "./features/reducer-factories/entity-reducer/entity-actions";
+export * from "./features/reducer-factories/entity-reducer/entity-action-creators";
+export * from "./features/reducer-factories/entity-reducer/entity-selectors";
 
-export * from "./features/reducer-factories-with-reducer-id/entity-crud-status-reducer/entity-crud-status-redux";
+export * from "./features/reducer-factories/entity-crud-status-reducer/entity-crud-status-redux";
 
-export * from "./features/reducer-factories-with-reducer-id/modified-field-reducer/modified-field-redux";
+export * from "./features/reducer-factories/modified-field-reducer/modified-field-redux";
 
-export * from "./features/reducer-factories-with-reducer-id/selected-ids-reducer/selected-ids-reducer";
-export * from "./features/reducer-factories-with-reducer-id/selected-ids-reducer/selected-ids-actions";
-export * from "./features/reducer-factories-with-reducer-id/selected-ids-reducer/selected-ids-selectors";
+export * from "./features/reducer-factories/selected-ids-reducer/selected-ids-reducer";
+export * from "./features/reducer-factories/selected-ids-reducer/selected-ids-actions";
+export * from "./features/reducer-factories/selected-ids-reducer/selected-ids-action-creators";
+export * from "./features/reducer-factories/selected-ids-reducer/selected-ids-selectors";
 
-export * from "./features/reducer-factories-with-reducer-id/sort-order-by-id-reducer/sort-order-by-id-reducer";
-export * from "./features/reducer-factories-with-reducer-id/sort-order-by-id-reducer/sort-order-by-id-actions";
-export * from "./features/reducer-factories-with-reducer-id/sort-order-by-id-reducer/sort-order-by-id-selectors";
+export * from "./features/reducer-factories/sort-order-by-id-reducer/sort-order-by-id-reducer";
+export * from "./features/reducer-factories/sort-order-by-id-reducer/sort-order-by-id-actions";
+export * from "./features/reducer-factories/sort-order-by-id-reducer/sort-order-by-id-action-creators";
+export * from "./features/reducer-factories/sort-order-by-id-reducer/sort-order-by-id-selectors";
 
-export * from "./features/reducer-factories-with-reducer-id/sort-order-reducer/sort-order-reducer";
-export * from "./features/reducer-factories-with-reducer-id/sort-order-reducer/sort-order-actions";
-export * from "./features/reducer-factories-with-reducer-id/sort-order-reducer/sort-order-selectors";
+export * from "./features/reducer-factories/sort-order-reducer/sort-order-reducer";
+export * from "./features/reducer-factories/sort-order-reducer/sort-order-actions";
+export * from "./features/reducer-factories/sort-order-reducer/sort-order-action-creators";
+export * from "./features/reducer-factories/sort-order-reducer/sort-order-selectors";
